refonte des packages de test suppression du propertyValue avec la conf de prod ajouts des annotations de conf de test sur les class de tests
