+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/Application.java
@@ -5,10 +5,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.PropertySource;
 
 
-@PropertySource(value = "application-prod.properties")
 @SpringBootApplication
 public class Application {
