+++ Poseiden-skeleton/src/test/java/com/nnk/springboot/repositories/BidTests.java
@@ -1,0 +1,48 @@
+package com.nnk.springboot.repositories;
+
+import com.nnk.springboot.domain.BidList;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+import java.util.Optional;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ActiveProfiles("test")
+public class BidTests {
+
+	@Autowired
+	private BidListRepository bidListRepository;
+
+	@Test
+	public void bidListTest() {
+		BidList bid = new BidList("Account Test", "Type Test", 10d);
+
+		// Save
+		bid = bidListRepository.save(bid);
+		Assert.assertNotNull(bid.getBidListId());
+		Assert.assertEquals(bid.getBidQuantity(), 10d, 10d);
+
+
+		// Update
+		bid.setBidQuantity(20d);
+		bid = bidListRepository.save(bid);
+		Assert.assertEquals(bid.getBidQuantity(), 20d, 20d);
+
+		// Find
+		List<BidList> listResult = bidListRepository.findAll();
+		Assert.assertTrue(listResult.size() > 0);
+
+		// Delete
+		Integer id = bid.getBidListId();
+		bidListRepository.delete(bid);
+		Optional<BidList> bidList = bidListRepository.findById(id);
+		Assert.assertFalse(bidList.isPresent());
+	}
+}
